feat: 修改token相关逻辑，现在token包含的信息是userId而不是username

+++ src/main/java/org/seec/muggle/auror/dao/BaseOperation.java
@@ -30,7 +30,7 @@ public interface BaseOperation<T> {
      * @param id 记录id
      * @return 影响行数
      */
-    T get(Integer id);
+    T get(Long id);
 
     /**
      * 删除记录
@@ -38,5 +38,5 @@ public interface BaseOperation<T> {
      * @param id 记录id
      * @return 影响行数
      */
-    T delete(Integer id);
+    T delete(Long id);
 }
